Make the dashboard switcher non-sticky
Also only includes it in the dashboards, i.e. not on the settings
and FAQ pages.
