+++ frontend/src/components/layout/Layout.module.scss
@@ -151,7 +151,7 @@ $toastify-toast-close-button: var(--toast-close-button);
   // overlap the AppPicker drawer).
   // Thus, upping the z-index stack the header, and menus
   // expanding from it, on top again:
-  z-index: 3;
+  z-index: 2;
   display: grid;
   grid-template-columns: 1fr;
   grid-template-areas:
@@ -257,18 +257,6 @@ $toastify-toast-close-button: var(--toast-close-button);
   height: 100%;
   display: flex;
   flex-direction: column;
-  z-index: 2; //z-index is a unit higher, so that it sits above the mask/phone duo nav
-}
-
-.mobile-nav-duo-wrapper {
-  z-index: 1; //z-index is a unit lower, so that the mask/phone duo nav slides sits under the expanded mobile menu
-  position: sticky;
-  top: 0;
-  padding-bottom: $layout-xl; // Adds padding to make room for the mask/phone nav on mobile
-
-  @media screen and (min-width: $screen-md) {
-    padding-bottom: 0;
-  }
 }
 
 .content {
