+++ frontend/src/components/layout/navigation/DashboardSwitcher.module.scss
@@ -1,0 +1,62 @@
+@import "~@mozilla-protocol/core/protocol/css/includes/lib";
+@import "../../../styles/tokens.scss";
+
+.nav-mask-phone {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  border-bottom: 1px solid $color-light-gray-20;
+  background-color: white;
+
+  > * {
+    flex: 1 1 0;
+    padding: $spacing-lg 0 $spacing-md 0;
+    justify-content: center;
+    display: flex;
+    border-bottom: 5px solid transparent;
+
+    &.is-active {
+      border-bottom: 5px solid $color-purple-40;
+    }
+
+    &:hover {
+      background-color: $color-purple-05;
+    }
+  }
+
+  .nav-mask-phone-icon {
+    color: $color-grey-40;
+    stroke: $color-grey-40;
+    stroke-width: 0.5px;
+
+    p,
+    p.is-active {
+      color: $color-purple-70;
+      text-transform: uppercase;
+    }
+
+    &.is-active {
+      color: $color-purple-40;
+      stroke: $color-purple-40;
+    }
+  }
+
+  .phone-icon-new-wrapper {
+    display: flex;
+    flex-direction: row;
+    align-content: center;
+
+    p {
+      @include text-body-sm;
+      margin: 0;
+      font-weight: 600;
+      align-self: center;
+    }
+  }
+}
+
+.hidden-desktop {
+  @media screen and (min-width: $screen-md) {
+    display: none;
+  }
+}
